+++ lib/glimmer/libui.rb
@@ -23,7 +23,7 @@ require 'glimmer/fiddle_consumer'
 
 module Glimmer
   module LibUI
-    ICON = File.expand_path('../../icons/glimmer.png', __dir__)
+    ICON = File.expand_path('../../icons/blank.png', __dir__)
     class << self
       include Glimmer::FiddleConsumer