+++ lib/glimmer/libui/control_proxy/table_proxy.rb
@@ -50,20 +50,15 @@ module Glimmer
           @cell_rows = []
           @last_cell_rows ||= array_deep_clone(@cell_rows)
           @cell_rows_observer ||= Glimmer::DataBinding::Observer.proc do |new_cell_rows|
-            pd model; $stdout.flush
-            
             if model
               if @cell_rows.size < @last_cell_rows.size && @last_cell_rows.include_all?(*@cell_rows)
                 @last_cell_rows.array_diff_indexes(@cell_rows).reverse.each do |row|
                   ::LibUI.table_model_row_deleted(model, row)
-  pd 'deleted'; $stdout.flush
                   notify_custom_listeners('on_changed', row, :deleted, @last_cell_rows[row])
                 end
               elsif @cell_rows.size > @last_cell_rows.size && @cell_rows.include_all?(*@last_cell_rows)
                 @cell_rows.array_diff_indexes(@last_cell_rows).each do |row|
-  
                   ::LibUI.table_model_row_inserted(model, row)
-  pd 'inserted'; $stdout.flush
                   notify_custom_listeners('on_changed', row, :inserted, @cell_rows[row])
                 end
               else
@@ -165,8 +160,6 @@ module Glimmer
               model_attribute_observer.add_dependent(model_attribute_observer_registration => @model_attribute_array_observer_registration)
             end
             # TODO look if multiple notifications are happening as a result of observing array and observing model binding
-            pd property; $stdout.flush
-            pd new_value; $stdout.flush
             send("#{property}=", new_value) unless @last_cell_rows == new_value
           end
           model_attribute_observer_registration = model_attribute_observer.observe(model_binding)
@@ -218,7 +211,7 @@ module Glimmer
             when Column::TextColumnProxy, Column::ButtonColumnProxy, Column::TextColorColumnProxy, :text
               ::LibUI.new_table_value_string((expanded_cell_rows[row] && expanded_cell_rows[row][column]).to_s)
             when Column::ImageColumnProxy, Column::ImageTextColumnProxy, Column::ImageTextColorColumnProxy
-              if (OS.windows? && row == cell_rows.count) || (expanded_cell_rows[row][column].nil?)
+              if OS.windows? && row == cell_rows.count
                 img = Glimmer::LibUI::ICON
               else
                 img = expanded_cell_rows[row][column]
@@ -319,12 +312,6 @@ module Glimmer
           @libui = ControlProxy.new_control(@keyword, [@table_params])
           @libui.tap do
             @columns.each {|column| column.respond_to?(:build_control, true) && column.send(:build_control) }
-          end
-          if OS.windows?
-            pd 'add row'; $stdout.flush
-            cell_rows << @columns.map {nil}
-            pd 'delete added row'; $stdout.flush
-            cell_rows.pop
           end
         end